+++ src/test/java/com/imsweb/seerutilsgui/SeerProgressDialogTest.java
@@ -15,6 +15,7 @@ public class SeerProgressDialogTest {
 
     public static void main(String[] args) {
         SeerGuiUtils.setupGuiEnvForSeerProject();
+        SeerGuiUtils.setFontDelta(5);
         JFrame myFrame = new JFrame("SeerProgressDialogTest");
         myFrame.setLayout(new FlowLayout(FlowLayout.CENTER, 0, 55));
         myFrame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);